Render UI in hierarchical order
- We didn't need a tree structure in the end, as the iteration can
already be done with Transform system processHierarchy.
- One issue is that we don't have a single rootId which would be
a parent to all other UI entities. Maybe we should do that for
simplicity?
- Define viewports first, because 3D rendering can never be
under the UI, unless we make some render to texture, which seems
unnecessary.
- Fix togglebutton rendering.
- Oh yeah, Window release build crashes after previous commit,
because of exposing AssetSystem as a reference dependency in
UIScene. Works in debug. Works on macOS. Should fix later.