Remove superficial std::moves and move functions
- We had a bunch of move powered add functions for our components,
but those are completely useless, as std::vector will always copy
(when used with a value) to keep the data contiguous. So you
can't really std::move into a std::vector.
- Rename non-const versions of getters into modify functions.
This looks much better. Now you can see from the function call if
it might get modified or not.